refactor - deeds work
