+++ frontend/components/housing/DeanwoodDeeds.js
@@ -1,405 +1,15 @@
-import React, {useState} from "react";
-import MapDeanwoodDeeds from "../maps/MapDeanwoodDeeds";
+import React from "react";
+import {info} from "./deanwood_deeds_info";
+import Deeds from "./Deeds";
 
-const lot_58_square_5175 = <>
-    <table className="table">
-        <tbody>
-        <tr>
-            <th>Year</th>
-            <th>Price on deed</th>
-        </tr>
-        <tr>
-            <td>1926</td>
-            <td>$10</td>
-        </tr>
-        <tr>
-            <td>1932</td>
-            <td>$500</td>
-        </tr>
-        <tr>
-            <td>1942</td>
-            <td>$10</td>
-        </tr>
-        <tr>
-            <td>1956</td>
-            <td>$4400</td>
-        </tr>
-        <tr>
-            <td>1964</td>
-            <td>$5000</td>
-        </tr>
-        <tr>
-            <td>2004</td>
-            <td>$145,600</td>
-        </tr>
-        <tr>
-            <td>2013</td>
-            <td>$253,400</td>
-        </tr>
-        <tr>
-            <td>
-                <mark>2022 (estimate)</mark>
-            </td>
-            <td>
-                <mark>~$611,334</mark>
-            </td>
-        </tr>
-        </tbody>
-    </table>
-</>;
-const lot_89_square_5175 = <>
-    <table className="table">
-        <tbody>
-        <tr>
-            <th>Year</th>
-            <th>Price on deed</th>
-        </tr>
-        <tr>
-            <td>1944</td>
-            <td>$10</td>
-        </tr>
-        <tr>
-            <td>1969</td>
-            <td>$10</td>
-        </tr>
-        <tr>
-            <td>1987</td>
-            <td>$22500</td>
-        </tr>
-        <tr>
-            <td>2002</td>
-            <td>$134000</td>
-        </tr>
-        <tr>
-            <td>2006</td>
-            <td>$159000</td>
-        </tr>
-        <tr>
-            <td>2010</td>
-            <td>$10</td>
-        </tr>
-        <tr>
-            <td>2017</td>
-            <td>$410000</td>
-        </tr>
-        <tr>
-            <td>
-                <mark>2022 (estimate)</mark>
-            </td>
-            <td>
-                <mark>~$567,523</mark>
-            </td>
-        </tr>
-        </tbody>
-    </table>
-</>;
-const lot_63_square_5175 = <>
-    <table className="table">
-        <tbody>
-        <tr>
-            <th>Year</th>
-            <th>Price on deed</th>
-        </tr>
-        <tr>
-            <td>1947</td>
-            <td>$10</td>
-        </tr>
-        <tr>
-            <td>1953</td>
-            <td>$10</td>
-        </tr>
-        <tr>
-            <td>1964</td>
-            <td>$22500</td>
-        </tr>
-        <tr>
-            <td>1985</td>
-            <td>$10</td>
-        </tr>
-        <tr>
-            <td>1990</td>
-            <td>$100000</td>
-        </tr>
-        <tr>
-            <td>2009</td>
-            <td>$10</td>
-        </tr>
-        <tr>
-            <td>
-                <mark>2022 (estimate)</mark>
-            </td>
-            <td>
-                <mark>~$421,523</mark>
-            </td>
-        </tr>
-        </tbody>
-    </table>
-</>;
-const lot_54_square_5175 = <>
-    <table className="table">
-        <tbody>
-        <tr>
-            <th>Year</th>
-            <th>Price on deed</th>
-        </tr>
-        <tr>
-            <td>1942</td>
-            <td>$10</td>
-        </tr>
-        <tr>
-            <td>1987</td>
-            <td>$10</td>
-        </tr>
-        <tr>
-            <td>1985</td>
-            <td>$10</td>
-        </tr>
-        <tr>
-            <td>2007</td>
-            <td>$10</td>
-        </tr>
-        <tr>
-            <td>2019</td>
-            <td>$13000</td>
-        </tr>
-        <tr>
-            <td>2021</td>
-            <td>$565000</td>
-        </tr>
-        <tr>
-            <td>
-                <mark>2022 (estimate)</mark>
-            </td>
-            <td>
-                <mark>~$614,547</mark>
-            </td>
-        </tr>
-        </tbody>
-    </table>
-</>;
-const lot_90_square_5175 = <>
-    <table className="table">
-        <tbody>
-        <tr>
-            <th>Year</th>
-            <th>Price on deed</th>
-        </tr>
-        <tr>
-            <td>1944</td>
-            <td>$10</td>
-        </tr>
-        <tr>
-            <td>1968</td>
-            <td>$10</td>
-        </tr>
-        <tr>
-            <td>1969</td>
-            <td>$10</td>
-        </tr>
-        <tr>
-            <td>1983</td>
-            <td>$9000</td>
-        </tr>
-        <tr>
-            <td>2013</td>
-            <td>$0</td>
-        </tr>
-        <tr>
-            <td>
-                <mark>2022 (estimate)</mark>
-            </td>
-            <td>
-                <mark>~$218,820</mark>
-            </td>
-        </tr>
-        </tbody>
-    </table>
-</>;
-const lot_65_square_5175 = <>
-    <table className="table">
-        <tbody>
-        <tr>
-            <th>Year</th>
-            <th>Price on deed</th>
-        </tr>
-        <tr>
-            <td>1927</td>
-            <td>$10</td>
-        </tr>
-        <tr>
-            <td>1933</td>
-            <td>$10</td>
-        </tr>
-        <tr>
-            <td>1944</td>
-            <td>$10</td>
-        </tr>
-        <tr>
-            <td>1949</td>
-            <td>$10</td>
-        </tr>
-        <tr>
-            <td>1985</td>
-            <td>$10</td>
-        </tr>
-        <tr>
-            <td>1996</td>
-            <td>$0</td>
-        </tr>
-        <tr>
-            <td>2004</td>
-            <td>$0</td>
-        </tr>
-        <tr>
-            <td>2008</td>
-            <td>$10</td>
-        </tr>
-        <tr>
-            <td>
-                <mark>2022 (estimate)</mark>
-            </td>
-            <td>
-                <mark>~$218,820</mark>
-            </td>
-        </tr>
-        </tbody>
-    </table>
-</>;
-const lot_80_square_5175 = <>
-    <table className="table">
-        <tbody>
-        <tr>
-            <th>Year</th>
-            <th>Price on deed</th>
-        </tr>
-        <tr>
-            <td>1928</td>
-            <td>$10</td>
-        </tr>
-        <tr>
-            <td>1949</td>
-            <td>$10</td>
-        </tr>
-        <tr>
-            <td>1963</td>
-            <td>$10</td>
-        </tr>
-        <tr>
-            <td>2016</td>
-            <td>$10</td>
-        </tr>
-        <tr>
-            <td>
-                <mark>2022 (estimate)</mark>
-            </td>
-            <td>
-                <mark>~$408,533</mark>
-            </td>
-        </tr>
-        </tbody>
-    </table>
-</>;
-const lot_59_square_5175 = <>
-    <table className="table">
-        <tbody>
-        <tr>
-            <th>Year</th>
-            <th>Price on deed</th>
-        </tr>
-        <tr>
-            <td>1924</td>
-            <td>$0</td>
-        </tr>
-        <tr>
-            <td>1946</td>
-            <td>$10</td>
-        </tr>
-        <tr>
-            <td>1949</td>
-            <td>$10</td>
-        </tr>
-        <tr>
-            <td>1981</td>
-            <td>$0</td>
-        </tr>
-        <tr>
-            <td>
-                <mark>2022 (estimate)</mark>
-            </td>
-            <td>
-                <mark>~$408,533</mark>
-            </td>
-        </tr>
-        </tbody>
-    </table>
-</>;
+const tables = info.tables;
+const coordinates = info.coordinates;
 
-const coordinates = {
-    "lot_58_square_5175": [38.903752, -76.929477],
-    "lot_89_square_5175": [38.90446, -76.93054],
-    "lot_63_square_5175": [38.904446, -76.929479],
-    "lot_54_square_5175": [38.903724, -76.929066],
-    "lot_90_square_5175": [38.904477, -76.930442],
-    "lot_65_square_5175": [38.904467, -76.929209],
-    "lot_80_square_5175": [38.904465, -76.929834],
-    "lot_59_square_5175": [38.904046, -76.929446]
-};
-
-const tables = {
-    "lot_58_square_5175": lot_58_square_5175,
-    "lot_89_square_5175": lot_89_square_5175,
-    "lot_63_square_5175": lot_63_square_5175,
-    "lot_54_square_5175": lot_54_square_5175,
-    "lot_90_square_5175": lot_90_square_5175,
-    "lot_65_square_5175": lot_65_square_5175,
-    "lot_80_square_5175": lot_80_square_5175,
-    "lot_59_square_5175": lot_59_square_5175
-};
-
-const lots = Object.keys(tables);
-
-const fixTitle = (name) => {
-    return name.replaceAll("_", " ").toUpperCase();
-};
 export const DeanwoodDeeds = () => {
-    const [lotNum, setLotNum] = useState(0);
-    const [table, setTable] = useState(tables[lots[0]]);
-    const [title, setTitle] = useState(fixTitle(Object.keys(coordinates)[0]));
-    const selectMarker = (obj) => {
-        let selected = lots.filter((val) => {
-            return (coordinates[val][0] === obj.latlng.lat &&
-                coordinates[val][1] === obj.latlng.lng);
-        });
-        selected = selected[0];
-        setTable(tables[selected]);
-        setTitle(fixTitle(selected));
-        setLotNum(lots.indexOf(selected));
-    };
 
-
-    return (<>
-        <h2 className="p-0">
-            Deanwood Lots: value by year
-        </h2>
-        <div className="carousel">
-            <div className="p-lg-3 p-sm-2 container">
-                <div className="p-0 row">
-                    <div className="p-0 col-6">
-                        <h5 className="tab-header">
-                            {title}
-                        </h5>
-                        <MapDeanwoodDeeds selectMarker={selectMarker}
-                                          markers={Object.values(coordinates)}
-                                          selected={coordinates[lots[lotNum]]}/>
-                    </div>
-                    <div className="pl-2 col-6 table-container">
-                        {table}
-                    </div>
-                </div>
-            </div>
-        </div>
-
-    </>);
+    return (
+        <Deeds tables={tables} coordinates={coordinates}/>
+    );
 };
 
 export default DeanwoodDeeds;
